// src/types.ts
export interface LsrTableDefinition {
  id: string;
  name: string;
  columns: string[];
}